Initial implementation of composite negotiator with support for basic capabilities.
An initial implementation of a composite negotiator that can do the
negotiation for a list of basic capabilities. Basic capabilities are
capabilities which can only be turned on or off and do not require
additional conversation with the IRC server itself.

The implementation supports acked capabilities and will automatically
send the client ACK in case this is required by the IRC server.